+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog: HTTP Status
 
+## Trunk
+- update: change all of module
+- return http status and message
+
 ## Version 0.1.1
 - add: Travis CI